Add basic script for mirroring gitlab repos to github
This script will loop through the repos in the /wordpress-plugins group
on GitLab (will currently only loop through 2, whilst testing, but will
updates this to loop through all), and check if a corresponding repo
exists in the GitHub dxw-wordpress-plugins org.

If the repo does not exist on GitHub, it will create it and mirror from
the GitLab repo. If it does exist on GitHub, it will check if the sha
for the latest commit matches. If not, it will mirror the latest
contents of the repo from GitLab.

The intention is that a scheduled GitHub action will run this script,
action to be added in a later commit.