Allow for repos with space in name
Although most gitlab projects don't have a space in the "name" property
returned by the API, and are instead hyphen separated, some have a
spaced "name" property e.g. "Google Analytics". This causes the script
to fail where we use the name property to try and mirror the repo. The
"path" property is always hyphen-separated, and should work in all
cases.