Use GitHub API call to confirm if repo exists
Whilst using `git ls-remote` works locally, it is failing on GitHub, and
always returning an error. This means the script always thinks the repo
doesn't already exist, and then tries (and fails) to create it again.

I'm not entirely sure why this is, but I think it may be because we're
already inside a git repo (this one, checked out by the checkout
action), and that's causing a conflict.

So instead, let's call the API. If we get a 200 response, the repo
we're looking for exists, if we get anything else, it doesn't.

Because we just want the response code, we're using CURL to call the API
here, rather than `gh api`.