Use keep-workflow-alive to ensure scheduled workflow stays active
GitHub automatically disables scheduled workflows on public repos that
have had no activity for 60 days. It currently defines "activity" as git
activity on the default branch.

This causes an issue in this repo, as we need the task to mirror the
plugins to run indefinitely.

This commit adds dxw's fork of the keepalive-workflow action to address
this problem. This action adds an empty commit to the default branch
every 50 days, if there has been no other activity in the meantime. It's
not an ideal way of solving the problem, but it's better than the
workflow repeatedly deactivating, which is the only other option at the
moment.