Use v1.0.7 of keepalive-workflow
The previous version we were using (1.06) didn't push the commit it
generated to the repo, so therefore didn't generate the necessary
activity to stop the repo being marked as inactive. This latest version
fixes that issue (see
https://github.com/gautamkrishnar/keepalive-workflow/pull/3).

I've also switched to using the upstream version, rather than our fork.
If we pin to a specific commit we're happy with (as I've done here),
this is equally safe.