Tag latest commit with relevant abbreviated major version tag
This checks if the latest commit has an associated tag, and if so also
tags that commit with the beginning of that tag, up to the first `.`.
This is on the assumption that all the repos being mirrored have single
tags per commit, in the format `vx.y.z`, where the major version tag we
also want associated with the latest commit is `vx`.

This assumption should be correct for all plugins currently stored in
GitLab repos taken from the WordPress plugin directory using the process
on trifle.dxw.net.

If there is no tag associated with the latest commit, no additional tag
is added. This could be the case e.g. where a plugin was manually added
to GitLab. At some point we'll probably want to manually fix those
repos.