Mirror from wordpress.org not GitLab
Currently we maintain two plugin libraries: one
on GitHub and one on GitLab. The GitLab library
is mirrored from wordpress.org and the GitHub
mirror reflects the repositories in GitLab.

Eventually, we want to standardise on one library,
which will be in GitHub. To do this, we need to
ensure that both plugin libraries mirror
wordpress.org.

This commit adds a new Ruby script to mirror plugins
from wordpress.org. Specifically it:

* Uses the same algorithm we had in Whippet Racetrack
* Adds explicit error handling
* Adds more structured log output

The mirroring algorithm is roughly:

* For each plugin repository:

    * Use the wordpress.org REST API to find the latest
      version of the plugin as a semver string.
    * Use the GitHub REST API to find the latest tag
      in the mirror repo.
    * Compare the versions, if they are identical move
      to the next repository.
    * Clone the plugin repository from GitHub.
    * Download the latest version of the plugin as a
      .zip file from wordpress.org.
    * Unzip the plugin files _over_ the repo clone.
    * Add and commit the new files.
    * Tag the repo with the new version number and a
      major-mode version.
    * Push the new commits and the new tags back to
      the GitHub repo.

This is also roughly identical to the manual process we
use to update pro- and premium plugins.