Add debug code for GitHub testing
We want to run the wordpress mirror script for several
days before we can be confident to make it live. This commit
adds some debug code and comments out the `git push`
line, just to give us some extra information about
how the script performs.

This commit should be reverted when deprecate the old script.