Compare semver versions as objects
Previously to decide whether or not a repository needed
updating we compared version strings for equality. Some
plugins have a pro- version and a free version with
the same slug, so this had the effect that free plugin
version tags were being pushed _after_ premium versions.
For example, a version "3" tag would be pushed to a repo
when a version "4" tag already existed.

This commit compares semver versions as objects, not strings
and requires a strict greater-than relationship between
two versions for the script to push an update.