Workaround for incorrectly tagged releases
We have sometimes ended up with tags that start `vv` instead
of `v`. Probably due to `v...` tags on WordPress that then had
`v` prepended to them. Where this has happened once we have
deleted the tag, where it has happened on every release we
have not.

category-specific-rss-feed-menu is one example of a plugin
whose WordPress tags are prepended by a `v`.