Add global Git configuration to workflow
In the scheduled job we are seeing errors of the form:

    Author identity unknown

This doesn't appear in local dev envs, presumably
because developers all have Git fully configured.

This error also does not appear on the GitLab
scheduled jobs, possibly because we clone with
`--mirror` which sets up repositories slightly
differently to a normal clone operation.

This commit adds the small amount of configuration
that should be needed to commit, tag and push
from the mirrors.