Workaround for erroneous tags in old versions
Previously we made an assumption that tags from GitHub
repos could be sorted alphabetically.

This fails in the case of login-lockdown which has a
latest tag of v2.10 on both GitHub and GitLab, but also
has an old tag, vv1.8.1. In this case the script tries
to push a "new" tag of v2.10 to GitHub, which already
exists, leading to:

    fatal: tag 'v2.10' already exists

The cause of this is that we use `sort -V` to sort
semver versions. `sort` assumes that the input lines are
file names of the form PREFIX VERSION SUFFIX, where SUFFIX
matches the regular expression
"(.([A-Za-z~][A-Za-z0-9~]*)?)*", and in this case that
assumption fails.

In this commit we correct that by stripping out all
alphabetical characters before we sort and compare the
tags from the GitHub repo. This is the new algorithm:

      gh api ... --jq '.[].ref'  # Fetch all tags from GitHub
    | cut -d '/' -f 3            # refs/tags/v1.2.3 -> v1.2.3
    | grep -Eo "[0-9\.]+"        # Match only numbers and "."
    | sort -V                    # Sort semver versions
    | tail -n 1                  # Take the last version

This gives the correct result for login-lockdown:

    ==> login-lockdown is v2.10 on WordPress and v2.10 on GitHub
    ...which is already up to date with WordPress.org