Raise exception if Git cannot tag
Previously we made a conscious decision not to raise
exceptions for local Git operations. However, in practice
we have found some errors raised when trying to
create tags. Rather than let these proceed silently,
it would be easier to have them in the summary of
failed plugins at the end of the output.