Correctly catch errors from gh
Before we used Ruby exception handling to catch errors
from gh, when we get a list of repo tags from GitHub.

However, we are calling gh by running it as a subprocess
command, so it won't raise an exception in the normal
way. This means that repos that have no tags are not
getting a blank version string, as they should be.

This commit fixes that by using the `$CHILD_STATUS`
to check the exit code of the child process.