Skip plugins with 'skip-mirror' topic
Many of our plugins are premium plugins which may have
a related free plugins. Some free/premium plugins use
the same tag, which means that previous to this commit,
the mirror service would overwrite the current premium
plugin with the free version on wordpress.org.

This commit checks the topics for each repository and
if a topic called 'skip-mirror' exists, then no
attempt is made to find the equivalent slug on
wordpress.org.

A list of skipped plugins is printed at the end of
the log.

This topic has a number of uses, for example, we can
add it to this repository, and several in the
organisation which are not actually plugins, or where
we usually deploy from elsewhere.