Refactor PluginLibraryUpdater
This class was getting quite complicated and held two
separate responsibilities:

1. Generating JSON and creating Ruby objects from each
   JSON object.
2. Updating each plugin in turn and logging the output.

This commit creates a new class, PluginBuilder, which
takes on the first responsibility, leaving the
PluginLibraryUpdated to deal with updating the library.

This has also had the side-effect of splitting the Plugin
constructor into two parts, one which initialises
attributes that are known from the JSON and one new
method which fetches information from other APIs.

This is slightly less readable, but it does mean that
when we skip over plugins, we can avoid making some
API calls, and still have a Plugin object for the
updater to query.