Add WordPressPluginJSONContainsError exception
Some plugins return an HTTP error on their wordpress.org
API endpoint and some return a JSON object which contains
an error message, with a reason and sometimes a 'reason_text'
key.

Previously, we did not take care to differentiate between
these different errors. In future though, we want to be able
to add metadata to a repository to tell developers whether
a plugin has been withdrawn, or similar and why.

The reason why a plugin has been withdrawn are important
to note, as different reasons will require different
actions from developers. For example, 'merged into core'
is a different situation to 'security issue'.