Only allow the GitLab mirror action to run on workflow_dispatch
Before: the mirroring of plugins from the GitLab upstream ran on a
schedule

Now: we're switching to using the WordPress directory as upstream, so we
don't want to run this action on a schedule anymore (but we're not quite
ready to remove it). So we're switching to running on manual
workflow_dipsatch events only.