Allow re-tagging of current version on force update
Before: if the script was run with `FORCE_UPDATE` set to true, the
`commit_and_tag` method would fail when attempting to apply the tag for
the current version, because that tag already exists. At this point, the
update process for the plugin in question would stop, with the effect
that no major version tag would be applied, even if no major version tag
previously existed.

Now: if `FORCE_UPDATE` is set to true, we add `-f` to the tag command
for the current main version tag, to prevent this error. It would
probably be safe to always do this, regardless of whether `FORCE_UPDATE`
is set or not, but as a precaution I'm only running it when that is set
to true.