Add initial major version tagging script
Given that this org will now have premium plugins pushed directly to it
(rather than mirrored from GitHub), we want a way to automatically
generate the appropriate major version tag for those repos, rather than
having to maintain the major version tag manually.

Therefore, it makes sense to split the logic for generating major
version tags for the repos within this org out from the process that
mirrors them from WordPress, so it applies to all repos regardless of
source.

This is the initial version of that script. It's heavily based on the
implementation of the main mirroring script. There are a few things
to note in particular here:

a) We look for any repos that have been updated within the last 12
hours, and apply major version tags to those. We haven't yet settled on
exactly what scheduled this script will run, but it'll probably be more
frequest that the script that does the actual mirroring, so this should
catch any updated repos
b) As yet we don't have any `force-update` functionality to make the
script re-tag all repos. That'll come in a later commit
c) For now, I've commented out the functionality that would actually
push the tagged commits, so the script should just report on what it
would do even if `dry_run` isn't set
d) We don't worry about identifying if a repo has already been updated.
There's not a simple way of getting the info from the API to do that,
and re-applying the same tag to the commit it already applies to doesn't
do any harm
e) We apply the major version tag to the same commit that has the full
version tag that the major tag is sourced from, rather than the HEAD
commit on the default branch. Therefore, if this and the mirroring
script should overlap (e.g. so that we identified the latest version as
being v6, but in the meantime the first commit of v7 is added to the
repo), we'll still tag the correct commit.