Ensure we always return an error if the version tag is empty
Before: the check for the success of the call to the gh api to get tags
references would always succeed in some environments, even if no tags
existed. This is because, depending on the config of the environment,
`$CHILD_STATUS` could return the exit code of the final command in the
piped sequence, rather than of any which errored (which the gh api could
would if no tags exist). This would result in repos which had no tags
returning a latest version of `v.`.

Now: as well as checking if the sequence of commands has exited
successfully, we check that the version tag returned does not equal `.`
(which occurs if the sequence of commands do no return an error exit
code, even if no tags actually exist). In this case, we still return an
empty version tag, which results in the Missing Version Error being
raised.