Do not set 'reason-' topics without a reason
The JSON we get from WordPress will offer a reason why
plugins have been closed or removed from the upstream
library. However, sometimes the reason is an empty string
and previously that led to a small number (24) plugins
in our library being given a 'reason-' topic.

This commit makes the relevant method a little more
defensive and does not set the topic if 'reason' is
an empty string in the JSON.

We also make the method a little more Ruby-ish by removing
some lexical scope with the 'return/unless' pattern.