FIX version discovery for empty GitHub repos
Previously, we assumed that if 'gh' and 'jq' could not find
any version information from a GitHub plugin repo then the
shell command would exit with a non-zero exit code.

In practice, we have found (with the 'loco-translate' plugin)
that the shell exec writes some text to stderr and exit
with status 0. However, it returns the string '.\n'.

In this commit we check for that string that is returned
by 'jq' and if that is the result of the shell exec we
return an empty string, indicating that the repo had no
tags.

This is important because when we mirror an entirely new
plugin, we do so from a completely empty repository.