Factor error classes into a mixin
We're expecting to add at least one new script to this
repo in the current PR and there is already quite a bit
of repeated code. Also, the mirror script is too long.

As a first step to help with this, this commit factors
new error classes into a mixin in the new 'lib/' directory.

Future PRs should try to factor out common code wherever
it is practical to do so.

This commit should not change the behaviour of either
of the top-level scripts.