Mirror cleans local files before pushing
We want to verify that the files we have committed to
a Git repo are the same files that we downloaded from
wordpress.org. If we leave behind any files that are
in the Git repo but not committed, this may lead to
our verification step giving incorrect results.

To avoid this situation, this commit runs a 'git clean'
on the repo after the files have been committed and
before they are pushed to the remote.

Since we run 'git add -A -f .' this step should really
not be necessary, but it may avoid a regression if we
change the commit_and_tag method in future.