Extract and lift plugin updater class
Plugin updaters iterate over a plugin collection, update each
plugin in some way and print a summary of their results.

This code is (mostly) repeated in the two main scripts we have
already, and so this commit extracts the common code into a
superclass which the two scripts then extend.