Extract and lift common plugin model code
Both our main scripts have a class which represents a plugin
in some way, and both classes include code for cloning
the related plugin repository on GitHub and finding the
latest version tag for that repo.

This commit factors that common code into an abstract
class. Not all of the lifted methods are shared between
both classes, some methods that are not dependent on the
particular algorithm from their source class have also
been lifted.