Checksum validator ignores paths from trunk/
This is a workaround for an issue in geo-my-wp v4.5.2.

For this plugin, a number of files which did not exist
in the repo were added to the checksum JSON. These file
paths started with the string "trunk/" so we believe this
was an error introduced when switching between Git (used
in the GitHub version of the repo) and SVN (used in wordpress.org).

Therefore, it seems safe to ignore errors from trunk/ paths.

    ==> Failed to update the following 1 plugin(s):

    *** geo-my-wp failed to update with message:
    trunk/assets/css/gmw.admin.css checksum mismatch. File does not exist in GitHub repository.