Skip mirroring pre-release versions of plugins
A pre-release version of a plugin is one with an
alphabetic character in the semver version, for example
v1.2.3-alpha. We do not wish to mirror these versions,
for two reasons. Firstly, they may not be stable releases.

More importantly, because a pre-release semver object cannot
be compared to a stable release version using the Gemfiles,
library. This leads to a bug where a stable release cannot
be mirrored because a pre-release version of a plugin
has previously been written into the library. In the
past we have temporarily fixed this by removing the
pre-release tag manually, but a this commit represents
a more permanent fix.