Send a Slack message when plugins cannot be mirrored
This commit adds a new GitHub action which sends a Slack
message detailing why mirror actions for specific plugins
failed, so that the Maintenance and Support team do not
have to manually check the logs.

This requires a new action secret 'SLACK_BAD_MIRROR_CHANNEL_ID'
which can be different from the channel ID that receives reports
that the whole mirror service has failed.